+++ src/components/PdfHighlighter.tsx
@@ -89,6 +89,9 @@ interface Props<T_HT> {
     categoryLabels: Array<{ label: string; background: string }>
   ) => JSX.Element | null;
   enableAreaSelection: (event: MouseEvent) => boolean;
+  getPageCount: (pageCount: number) => void;
+  getCurrentPage: (currentPage: number) => void;
+  destinationPage?: number;
 }
 
 const EMPTY_ID = "empty-id";
@@ -146,6 +149,9 @@ export class PdfHighlighter<T_HT extends IHighlight> extends PureComponent<
       const { ownerDocument: doc } = ref;
       eventBus.on("textlayerrendered", this.onTextLayerRendered);
       eventBus.on("pagesinit", this.onDocumentReady);
+      eventBus.on("pagechanging", () =>
+        this.props.getCurrentPage(this.viewer.currentPageNumber)
+      );
       doc.addEventListener("selectionchange", this.onSelectionChange);
       doc.addEventListener("keydown", this.handleKeyDown);
       doc.defaultView?.addEventListener("resize", this.debouncedScaleValue);
@@ -172,6 +178,10 @@ export class PdfHighlighter<T_HT extends IHighlight> extends PureComponent<
     }
     if (prevProps.highlights !== this.props.highlights) {
       this.renderHighlights(this.props);
+    }
+    const page = this.props.destinationPage;
+    if (page && prevProps.destinationPage !== page) {
+      this.goToPage(page);
     }
   }
 
@@ -213,6 +223,13 @@ export class PdfHighlighter<T_HT extends IHighlight> extends PureComponent<
       textLayer.div,
       "PdfHighlighter__highlight-layer"
     );
+  }
+
+  goToPage(page: number) {
+    if (page < 1 || page > this.viewer.pagesCount) {
+      return;
+    }
+    this.viewer.currentPageNumber = page;
   }
 
   groupHighlightsByPage(highlights: Array<T_HT>): {
@@ -485,6 +502,8 @@ export class PdfHighlighter<T_HT extends IHighlight> extends PureComponent<
     this.handleScaleValue();
 
     scrollRef(this.scrollTo);
+
+    this.props.getPageCount(this.viewer.pagesCount);
   };
 
   onSelectionChange = () => {